Separate DB worker from sync
